Performance improvement.
Seperate functions were more effective than closures

Avoiding declaring res also increased performance

rebased

Added contributor to package.json
